modal_worker/requirements.txt
faster-whisper
pyannote.audio
boto3
textblob
modal
